horizontal movement corrected

+++ src/game/consts.rs
@@ -1,2 +1,2 @@
-pub const BOARD_HEIGHT: f32 = 20.;
-pub const BOARD_WIGTH: f32 = 10.;
+pub const BOARD_HEIGHT: i32 = 20;
+pub const BOARD_WIGTH: i32 = 10;
